incremental changes 05/22/2025
