+++ ush/global_chem/global_chem_atmnos_plots_performance_diagram.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 '''
-Name: global_ens_chem_plots_performance_diagram.py
+Name: global_chem_atmnos_plots_performance_diagram.py
 Original Author: Mallory Row (mallory.row@noaa.gov)
 Contact(s): Ho-Chun Huang (ho-chun.huang@noaa.gov)
 Abstract: This script generates a performance_diagram plot.
@@ -20,8 +20,8 @@ import matplotlib
 matplotlib.use('agg')
 import matplotlib.pyplot as plt
 import matplotlib.dates as md
-import global_ens_chem_util as gda_util
-from global_ens_chem_plots_specs import PlotSpecs
+import global_chem_atmnos_util as gda_util
+from global_chem_atmnos_plots_specs import PlotSpecs
 
 class PerformanceDiagram:
     """
