+++ ush/global_chem/global_chem_atmnos_plots_valid_hour_average.py
@@ -1,11 +1,11 @@
 #!/usr/bin/env python3
 '''
-Name: global_ens_chem_plots_threshold_average.py
+Name: global_chem_atmnos_plots_valid_hour_average.py
 Original Author: Mallory Row (mallory.row@noaa.gov)
 Contact(s): Ho-Chun Huang (ho-chun.huang@noaa.gov)
-Abstract: This script generates a threshold average plot.
-          (x-axis: threshold value; y-axis: statistics value)
-          (EVS Graphics Naming Convention: threshmean)
+Abstract: This script generates a valid hour average plot.
+          (x-axis: valid hour; y-axis: statistics value)
+          (EVS Graphics Naming Convention: vhrmean)
 '''
 
 import sys
@@ -21,18 +21,17 @@ import matplotlib
 matplotlib.use('agg')
 import matplotlib.pyplot as plt
 import matplotlib.dates as md
-import re
-import global_ens_chem_util as gda_util
-from global_ens_chem_plots_specs import PlotSpecs
+import global_chem_atmnos_util as gda_util
+from global_chem_atmnos_plots_specs import PlotSpecs
 
-class ThresholdAverage:
+class ValidHourAverage:
     """
-    Make a threshold average graphic
+    Make a valid hour average graphic
     """
 
     def __init__(self, logger, input_dir, output_dir, model_info_dict,
                  date_info_dict, plot_info_dict, met_info_dict, logo_dir):
-        """! Initialize ThresholdAverage class
+        """! Initialize ValidHourAverage class
 
              Args:
                  logger          - logger object
@@ -55,14 +54,14 @@ class ThresholdAverage:
         self.met_info_dict = met_info_dict
         self.logo_dir = logo_dir
 
-    def make_threshold_average(self):
-        """! Make the threshold average graphic
+    def make_valid_hour_average(self):
+        """! Make the valid hour average graphic
 
              Args:
 
              Returns:
         """
-        self.logger.info(f"Plot Type: Threshold Average")
+        self.logger.info(f"Plot Type: Valid Hour Average")
         self.logger.debug(f"Input directory: {self.input_dir}")
         self.logger.debug(f"Output directory: {self.output_dir}")
         self.logger.debug(f"Model information dictionary: "
@@ -73,84 +72,104 @@ class ThresholdAverage:
                           +f"{self.plot_info_dict}")
         # Check stat
         if self.plot_info_dict['stat'] == 'FBAR_OBAR':
-            self.logger.error("Cannot make threshold_average for stat "
+            self.logger.error("Cannot make valid_hour_average for stat "
                               +f"{self.plot_info_dict['stat']}")
             sys.exit(1)
-        # Get dates to plot
-        self.logger.debug("Making valid and init date arrays")
-        valid_dates, init_dates = gda_util.get_plot_dates(
-            self.logger,
-            self.date_info_dict['date_type'],
-            self.date_info_dict['start_date'],
-            self.date_info_dict['end_date'],
-            self.date_info_dict['valid_hr_start'],
-            self.date_info_dict['valid_hr_end'],
-            self.date_info_dict['valid_hr_inc'],
-            self.date_info_dict['init_hr_start'],
-            self.date_info_dict['init_hr_end'],
-            self.date_info_dict['init_hr_inc'],
-            self.date_info_dict['forecast_hour']
-        )
-        format_valid_dates = [valid_dates[d].strftime('%Y%m%d_%H%M%S') \
-                              for d in range(len(valid_dates))]
-        format_init_dates = [init_dates[d].strftime('%Y%m%d_%H%M%S') \
-                             for d in range(len(init_dates))]
-        if self.date_info_dict['date_type'] == 'VALID':
-            self.logger.debug("Based on date information, plot will display "
-                              +"valid dates "+', '.join(format_valid_dates)+" "
-                              +"for forecast hour "
-                              +f"{self.date_info_dict['forecast_hour']} "
-                              +"with initialization dates "
-                              +', '.join(format_init_dates))
-            plot_dates = valid_dates
-        elif self.date_info_dict['date_type'] == 'INIT':
-            self.logger.debug("Based on date information, plot will display "
-                              +"initialization dates "
-                              +', '.join(format_init_dates)+" "
-                              +"for forecast hour "
-                              +f"{self.date_info_dict['forecast_hour']} "
-                              +"with valid dates "
-                              +', '.join(format_valid_dates))
-            plot_dates = init_dates
-        # Make dataframe for all thresholds
-        self.logger.info(f"Reading in model stat files from {self.input_dir}")
-        self.logger.info("Building dataframe for all thresholds")
-        # Make dataframe for all thresholds
+        # Make dataframe for all valid hour
+        self.logger.info("Building dataframe for all valid hours")
         fcst_units = []
-        for fcst_var_thresh in self.plot_info_dict['fcst_var_threshs']:
-            self.logger.debug("Building data for forecast threshold "
-                              +f"{fcst_var_thresh}")
-            fcst_var_thresh_idx = (self.plot_info_dict['fcst_var_threshs']\
-                                   .index(fcst_var_thresh))
-            obs_var_thresh = (self.plot_info_dict['obs_var_threshs']\
-                              [fcst_var_thresh_idx])
-            all_model_df = gda_util.build_df(
-                'make_plots', self.logger, self.input_dir, self.output_dir,
-                self.model_info_dict, self.met_info_dict,
-                self.plot_info_dict['fcst_var_name'],
-                self.plot_info_dict['fcst_var_level'],
-                fcst_var_thresh,
-                self.plot_info_dict['obs_var_name'],
-                self.plot_info_dict['obs_var_level'],
-                obs_var_thresh,
-                self.plot_info_dict['line_type'],
-                self.plot_info_dict['grid'],
-                self.plot_info_dict['vx_mask'],
-                self.plot_info_dict['interp_method'],
-                self.plot_info_dict['interp_points'],
-                self.date_info_dict['date_type'],
-                plot_dates, format_valid_dates,
-                self.date_info_dict['forecast_hour']
-            )
-            fcst_units.extend(
-                all_model_df['FCST_UNITS'].values.astype('str').tolist()
-            )
+        valid_hours = np.arange(
+            int(self.date_info_dict['valid_hr_start']),
+            int(self.date_info_dict['valid_hr_end'])\
+            +int(self.date_info_dict['valid_hr_inc']),
+            int(self.date_info_dict['valid_hr_inc'])
+        )
+        self.logger.info("Reading in model stat files "
+                         +f"from {self.input_dir}")
+        for valid_hour in valid_hours:
+            self.logger.debug(f"Building data for valid hour {valid_hour}")
+            for forecast_hour in self.date_info_dict['forecast_hours']:
+                self.logger.debug(f"Building data for forecast hour {forecast_hour}")
+                # Get dates to plot
+                self.logger.debug("Making valid and init date arrays")
+                valid_dates, init_dates = gda_util.get_plot_dates(
+                    self.logger,
+                    self.date_info_dict['date_type'],
+                    self.date_info_dict['start_date'],
+                    self.date_info_dict['end_date'],
+                    str(valid_hour),
+                    str(valid_hour),
+                    '24',
+                    self.date_info_dict['init_hr_start'],
+                    self.date_info_dict['init_hr_end'],
+                    self.date_info_dict['init_hr_inc'],
+                    forecast_hour
+                )
+                format_valid_dates = [valid_dates[d].strftime('%Y%m%d_%H%M%S') \
+                                      for d in range(len(valid_dates))]
+                format_init_dates = [init_dates[d].strftime('%Y%m%d_%H%M%S') \
+                                     for d in range(len(init_dates))]
+                if self.date_info_dict['date_type'] == 'VALID':
+                    self.logger.debug("Based on date information, "
+                                      +"plot will display valid dates "
+                                      +', '.join(format_valid_dates)+" "
+                                      +"for forecast hour "
+                                      +f"{forecast_hour} with "
+                                      +"initialization dates "
+                                      +', '.join(format_init_dates))
+                    plot_dates = valid_dates
+                elif self.date_info_dict['date_type'] == 'INIT':
+                    self.logger.debug("Based on date information, "
+                                      +"plot will display "
+                                      +"initialization dates "
+                                      +', '.join(format_init_dates)+" "
+                                      +"for forecast hour "
+                                      +f"{forecast_hour} with valid dates "
+                                      +', '.join(format_valid_dates))
+                    plot_dates = init_dates
+                # Read in data
+                all_model_df = gda_util.build_df(
+                    'make_plots', self.logger, self.input_dir, self.output_dir,
+                    self.model_info_dict, self.met_info_dict,
+                    self.plot_info_dict['fcst_var_name'],
+                    self.plot_info_dict['fcst_var_level'],
+                    self.plot_info_dict['fcst_var_thresh'],
+                    self.plot_info_dict['obs_var_name'],
+                    self.plot_info_dict['obs_var_level'],
+                    self.plot_info_dict['obs_var_thresh'],
+                    self.plot_info_dict['line_type'],
+                    self.plot_info_dict['grid'],
+                    self.plot_info_dict['vx_mask'],
+                    self.plot_info_dict['interp_method'],
+                    self.plot_info_dict['interp_points'],
+                    self.date_info_dict['date_type'],
+                    plot_dates, format_valid_dates,
+                    str(forecast_hour)
+                )
+                fcst_units.extend(
+                    all_model_df['FCST_UNITS'].values.astype('str').tolist()
+                )
+                model_idx_list = (
+                    all_model_df.index.get_level_values(0).unique().tolist()
+                )
+                all_model_df.rename(
+                    index=lambda s: s+'f'+str(forecast_hour).zfill(3),
+                    level=1, inplace=True
+                )
+                if forecast_hour \
+                        == self.date_info_dict['forecast_hours'][0]:
+                    all_forecast_hour_all_model_df = all_model_df
+                else:
+                    all_forecast_hour_all_model_df = pd.concat(
+                        [all_forecast_hour_all_model_df,all_model_df]
+                    )
             # Calculate statistic mean and 95% confidence intervals
             self.logger.info(f"Calculating statstic {self.plot_info_dict['stat']} "
                              +f"from line type {self.plot_info_dict['line_type']} "
                              +"average and 95% confidence intervals")
             stat_df, stat_array = gda_util.calculate_stat(
-                self.logger, all_model_df, self.plot_info_dict['line_type'],
+                self.logger, all_forecast_hour_all_model_df,
+                self.plot_info_dict['line_type'],
                 self.plot_info_dict['stat']
             )
             model_idx_list = (
@@ -164,18 +183,16 @@ class ThresholdAverage:
                 for model_idx in model_idx_list:
                     model_idx_num = model_idx_list.index(model_idx)
                     stat_df.loc[model_idx] = stat_array[model_idx_num,:]
-                    all_model_df.loc[model_idx] = (
-                        all_model_df.loc[model_idx].where(
+                    all_forecast_hour_all_model_df.loc[model_idx] = (
+                        all_forecast_hour_all_model_df.loc[model_idx].where(
                             stat_df.loc[model_idx].notna()
                     ).values)
-            if fcst_var_thresh == self.plot_info_dict['fcst_var_threshs'][0]:
-                threshs_avg_df = pd.DataFrame(
-                    np.nan, model_idx_list,
-                    columns=self.plot_info_dict['fcst_var_threshs']
+            if valid_hour == valid_hours[0]:
+                valid_hours_avg_df = pd.DataFrame(
+                    np.nan, model_idx_list, columns=valid_hours
                 )
-                threshs_ci_df = pd.DataFrame(
-                    np.nan, model_idx_list,
-                    columns=self.plot_info_dict['fcst_var_threshs']
+                valid_hours_ci_df = pd.DataFrame(
+                    np.nan, model_idx_list, columns=valid_hours
                 )
             for model_idx in model_idx_list:
                 model_idx_num = model_idx_list.index(model_idx)
@@ -186,15 +203,15 @@ class ThresholdAverage:
                     calc_avg_df = stat_df.loc[model_idx]
                 else:
                     avg_method = 'aggregation'
-                    calc_avg_df = all_model_df.loc[model_idx]
-                model_idx_thresh_avg = gda_util.calculate_average(
+                    calc_avg_df = all_forecast_hour_all_model_df.loc[model_idx]
+                model_idx_valid_hour_avg = gda_util.calculate_average(
                     self.logger, avg_method, self.plot_info_dict['line_type'],
                     self.plot_info_dict['stat'], calc_avg_df
                 )
-                if not np.isnan(model_idx_thresh_avg) \
-                        and not np.ma.is_masked(model_idx_thresh_avg):
-                    threshs_avg_df.loc[model_idx, fcst_var_thresh] = (
-                        model_idx_thresh_avg
+                if not np.isnan(model_idx_valid_hour_avg) \
+                        and not np.ma.is_masked(model_idx_valid_hour_avg):
+                    valid_hours_avg_df.loc[model_idx, valid_hour] = (
+                        model_idx_valid_hour_avg
                     )
                 if model_idx == model_idx_list[0]:
                     model1_stat_df = stat_df.loc[model_idx]
@@ -206,6 +223,12 @@ class ThresholdAverage:
                                 -np.ma.count_masked(model_idx_model1_diff))
                     model_idx_model1_diff_mean = np.mean(model_idx_model1_diff)
                     model_idx_model1_diff_std = np.std(model_idx_model1_diff)
+                    ##Null Hypothesis: mean(M1-M2)=0,
+                    ##M1-M2 follows normal distribution.
+                    ##plot the 5% conf interval of difference of means
+                    ##F*SD/sqrt(N-1),
+                    ##F=1.96 for infinite samples, F=2.0 for nsz=60,
+                    ##F=2.042 for nsz=30, F=2.228 for nsz=10
                     if nsamples > 1:
                         model_idx_model1_diff_mean_std_err = (
                             model_idx_model1_diff_std/np.sqrt(nsamples-1)
@@ -220,25 +243,30 @@ class ThresholdAverage:
                             ci = 2.228 * model_idx_model1_diff_mean_std_err
                     else:
                         ci = np.nan
-                    threshs_ci_df.loc[model_idx, fcst_var_thresh] = ci
+                    valid_hours_ci_df.loc[model_idx, valid_hour] = ci
+                    #from scipy import stats
+                    #scipy_ci = stats.t.interval(
+                    #    alpha=0.95,
+                    #    df=len(np.ma.compressed(model_idx_model1_diff))-1,
+                    #    loc=0,
+                    #    scale=stats.sem(np.ma.compressed(model_idx_model1_diff))
+                    #)
         # Set up plot
         self.logger.info(f"Setting up plot")
-        plot_specs_ta = PlotSpecs(self.logger, 'threshold_average')
-        plot_specs_ta.set_up_plot()
-        n_xticks = 7
-        xticks = np.arange(0, len(self.plot_info_dict['fcst_var_threshs']),
-                           1)
-        if len(xticks) < n_xticks:
+        plot_specs_vha = PlotSpecs(self.logger, 'valid_hour_average')
+        plot_specs_vha.set_up_plot()
+        n_xticks = 8
+        if len(valid_hours) < n_xticks:
             xtick_intvl = 1
         else:
-            xtick_intvl = int(len(xticks)/n_xticks)
+            xtick_intvl = int(len(valid_hours)/n_xticks)
         stat_min_max_dict = {
             'ax1_stat_min': np.ma.masked_invalid(np.nan),
             'ax1_stat_max': np.ma.masked_invalid(np.nan),
             'ax2_stat_min': np.ma.masked_invalid(np.nan),
             'ax2_stat_max': np.ma.masked_invalid(np.nan)
         }
-        stat_plot_name = plot_specs_ta.get_stat_plot_name(
+        stat_plot_name = plot_specs_vha.get_stat_plot_name(
              self.plot_info_dict['stat']
         )
         fcst_units = np.unique(fcst_units)
@@ -249,7 +277,7 @@ class ThresholdAverage:
         elif len(fcst_units) == 0:
             self.logger.debug("Cannot get variables units, leaving blank")
             fcst_units = ['']
-        plot_title = plot_specs_ta.get_plot_title(
+        plot_title = plot_specs_vha.get_plot_title(
             self.plot_info_dict, self.date_info_dict,
             fcst_units[0]
         )
@@ -260,9 +288,9 @@ class ThresholdAverage:
                 plot_left_logo_path
             )
             left_logo_xpixel_loc, left_logo_ypixel_loc, left_logo_alpha = (
-                plot_specs_ta.get_logo_location(
-                    'left', plot_specs_ta.fig_size[0],
-                    plot_specs_ta.fig_size[1], plt.rcParams['figure.dpi']
+                plot_specs_vha.get_logo_location(
+                    'left', plot_specs_vha.fig_size[0],
+                    plot_specs_vha.fig_size[1], plt.rcParams['figure.dpi']
                 )
             )
         else:
@@ -275,61 +303,30 @@ class ThresholdAverage:
                 plot_right_logo_path
             )
             right_logo_xpixel_loc, right_logo_ypixel_loc, right_logo_alpha = (
-                plot_specs_ta.get_logo_location(
-                    'right', plot_specs_ta.fig_size[0],
-                    plot_specs_ta.fig_size[1], plt.rcParams['figure.dpi']
+                plot_specs_vha.get_logo_location(
+                    'right', plot_specs_vha.fig_size[0],
+                    plot_specs_vha.fig_size[1], plt.rcParams['figure.dpi']
                 )
             )
         else:
             plot_right_logo = False
             self.logger.debug(f"{plot_right_logo_path} does not exist")
-        image_name = plot_specs_ta.get_savefig_name(
+        image_name = plot_specs_vha.get_savefig_name(
             self.output_dir, self.plot_info_dict, self.date_info_dict
         )
         # Make plot
         self.logger.info(f"Making plot")
         fig, (ax1, ax2) = plt.subplots(2,1,
-                                       figsize=(plot_specs_ta.fig_size[0],
-                                                plot_specs_ta.fig_size[1]),
+                                       figsize=(plot_specs_vha.fig_size[0],
+                                                plot_specs_vha.fig_size[1]),
                                        sharex=True)
-        if self.plot_info_dict['fcst_var_name'] == 'DPT' \
-                and self.plot_info_dict['fcst_var_level'] == 'Z2':
-            plot_title = plot_title.replace('2 meter Dewpoint (K)',
-                                            '2 meter Dewpoint (F)')
         fig.suptitle(plot_title)
         ax1.grid(True)
         ax1.set_ylabel(stat_plot_name)
         ax2.grid(True)
-        ax2.set_xlabel('Threshold')
-        ax2.set_xlim([xticks[0], xticks[-1]])
-        ax2.set_xticks(xticks[::xtick_intvl])
-        if self.plot_info_dict['fcst_var_name'] == 'DPT' \
-                and self.plot_info_dict['fcst_var_level'] == 'Z2':
-            convert_thresh_list = []
-            for thresh in self.plot_info_dict['fcst_var_threshs']:
-                convert_thresh_K_to_F = round(
-                    round((((float(thresh[2:])-273.15)*9)/5)+32)
-                )
-                convert_thresh_list.append(
-                    f"{thresh[0:2]}{str(convert_thresh_K_to_F)}"
-                )
-            ax2.set_xticklabels(convert_thresh_list[::xtick_intvl])
-        elif self.plot_info_dict['fcst_var_name'] == 'AOTK':
-            convert_thresh_list = []
-            for thresh in self.plot_info_dict['fcst_var_threshs']:
-                convert_thresh_sign = thresh.replace("ge","$\u2265$")
-                convert_thresh_list.append(convert_thresh_sign)
-            ax2.set_xticklabels(convert_thresh_list[::xtick_intvl])
-        elif self.plot_info_dict['fcst_var_name'] == 'PMTF':
-            convert_thresh_list = []
-            for thresh in self.plot_info_dict['fcst_var_threshs']:
-                convert_thresh_sign = thresh.replace("gt","$\u003E$")
-                convert_thresh_list.append(convert_thresh_sign)
-            ax2.set_xticklabels(convert_thresh_list[::xtick_intvl])
-        else:
-            ax2.set_xticklabels(
-                self.plot_info_dict['fcst_var_threshs'][::xtick_intvl]
-            )
+        ax2.set_xlabel('Valid Hour')
+        ax2.set_xlim([valid_hours[0], valid_hours[-1]])
+        ax2.set_xticks(valid_hours[::xtick_intvl])
         ax2.set_ylabel('Difference')
         ax2.set_title('Difference from '
                       +self.model_info_dict['model1']['plot_name'], loc='left')
@@ -354,17 +351,15 @@ class ThresholdAverage:
                 right_logo_img_array, right_logo_xpixel_loc,
                 right_logo_ypixel_loc, zorder=1, alpha=right_logo_alpha
             )
-        model_plot_settings_dict = plot_specs_ta.get_model_plot_settings()
+        model_plot_settings_dict = plot_specs_vha.get_model_plot_settings()
         model_idx_list = (
-            threshs_avg_df.index.get_level_values(0).unique().tolist()
+            valid_hours_avg_df.index.get_level_values(0).unique().tolist()
         )
         ci_bar_max_widths = np.append(
-            np.diff(xticks),
-            xticks[-1]-xticks[-2]
+            np.diff(valid_hours), valid_hours[-1]-valid_hours[-2]
         )/1.5
         ci_bar_min_widths = np.append(
-            np.diff(xticks),
-            xticks[-1]-xticks[-2]
+            np.diff(valid_hours), valid_hours[-1]-valid_hours[-2]
         )/len(list(self.model_info_dict.keys()))
         ci_bar_intvl_widths = (
             (ci_bar_max_widths-ci_bar_min_widths)
@@ -375,7 +370,7 @@ class ThresholdAverage:
             model_num_name = model_idx.split('/')[1]
             model_num_plot_name = model_idx.split('/')[2]
             model_num_obs_name = self.model_info_dict[model_num]['obs_name']
-            model_num_data = threshs_avg_df.loc[model_idx]
+            model_num_data = valid_hours_avg_df.loc[model_idx]
             if model_num_name in list(model_plot_settings_dict.keys()):
                 model_num_plot_settings_dict = (
                     model_plot_settings_dict[model_num_name]
@@ -384,6 +379,8 @@ class ThresholdAverage:
                 model_num_plot_settings_dict = (
                     model_plot_settings_dict[model_num]
                 )
+            self.logger.debug(f"Plotting {model_num} [{model_num_name},"
+                              +f"{model_num_plot_name}]")
             masked_model_num_data = np.ma.masked_invalid(model_num_data)
             if model_num == 'model1':
                  model1_masked_model_num_data = masked_model_num_data
@@ -391,20 +388,13 @@ class ThresholdAverage:
                 len(masked_model_num_data)
                 - np.ma.count_masked(masked_model_num_data)
             )
-            thresh_values = np.asarray(
-                [x for x in \
-                         range(0,len(threshs_avg_df.columns.values.tolist()))],
-                dtype=float
-            )
-            masked_thresh_values = np.ma.masked_where(
+            masked_valid_hours = np.ma.masked_where(
                 np.ma.getmask(masked_model_num_data),
-                thresh_values
+                valid_hours_avg_df.columns.values.tolist()
             )
             if model_num_npts != 0:
-                self.logger.debug(f"Plotting {model_num} [{model_num_name},"
-                                  +f"{model_num_plot_name}]")
                 ax1.plot(
-                    np.ma.compressed(masked_thresh_values),
+                    np.ma.compressed(masked_valid_hours),
                     np.ma.compressed(masked_model_num_data),
                     color = model_num_plot_settings_dict['color'],
                     linestyle = model_num_plot_settings_dict['linestyle'],
@@ -437,19 +427,19 @@ class ThresholdAverage:
                 len(masked_model_num_model1_diff_data)
                 - np.ma.count_masked(masked_model_num_model1_diff_data)
             )
-            masked_diff_thresh_values = np.ma.masked_where(
+            masked_diff_valid_hours = np.ma.masked_where(
                 np.ma.getmask(masked_model_num_model1_diff_data),
-                thresh_values
+                valid_hours_avg_df.columns.values.tolist()
             )
             if model_num_diff_npts != 0:
                 self.logger.debug(f"Plotting {model_num} [{model_num_name},"
                                   +f"{model_num_plot_name}] difference from "
-                                  +f"model1 ["
+                                  +"model1 ["
                                   +f"{self.model_info_dict['model1']['name']},"
                                   +self.model_info_dict['model1']['plot_name']
                                   +"]")
                 ax2.plot(
-                    np.ma.compressed(masked_diff_thresh_values),
+                    np.ma.compressed(masked_diff_valid_hours),
                     np.ma.compressed(masked_model_num_model1_diff_data),
                     color = model_num_plot_settings_dict['color'],
                     linestyle = model_num_plot_settings_dict['linestyle'],
@@ -474,14 +464,14 @@ class ThresholdAverage:
             else:
                 self.logger.debug(f"{model_num} [{model_num_name},"
                                   +f"{model_num_plot_name}] difference from "
-                                  +f"model1 ["
+                                  +"model1 ["
                                   +f"{self.model_info_dict['model1']['name']},"
                                   +self.model_info_dict['model1']['plot_name']
                                   +"] has no points")
             if model_num == 'model1':
                 ax2.plot(
-                    thresh_values,
-                    np.zeros_like(thresh_values),
+                    valid_hours_avg_df.columns.values.tolist(),
+                    np.zeros_like(valid_hours_avg_df.columns.values.tolist()),
                     color = model_num_plot_settings_dict['color'],
                     linestyle = model_num_plot_settings_dict['linestyle'],
                     linewidth = model_num_plot_settings_dict['linewidth'],
@@ -492,21 +482,22 @@ class ThresholdAverage:
                 )
             if model_num != 'model1':
                 masked_model_num_model1_diff_ci_data = np.ma.masked_invalid(
-                    threshs_ci_df.loc[model_idx]
+                    valid_hours_ci_df.loc[model_idx]
                 )
                 model_num_ci_npts = (
                     len(masked_model_num_model1_diff_ci_data)
                     - np.ma.count_masked(masked_model_num_model1_diff_ci_data)
                 )
-                masked_ci_thresh_values = np.ma.masked_where(
+                masked_ci_valid_hours = np.ma.masked_where(
                     np.ma.getmask(masked_model_num_model1_diff_ci_data),
-                    thresh_values
+                    valid_hours_ci_df.columns.values.tolist()
                 )
                 self.logger.debug(f"Plotting {model_num} ["
                                   +f"{model_num_name},"
                                   +f"{model_num_plot_name}] difference "
-                                  +"from model1 ["
-                                  +f"{self.model_info_dict['model1']['name']},"
+                                  +"from mode1 ["
+                                  +self.model_info_dict['model1']['name']
+                                  +","
                                   +self.model_info_dict['model1']['plot_name']
                                   +"] confidence intervals")
                 if model_num_ci_npts != 0:
@@ -520,8 +511,8 @@ class ThresholdAverage:
                             or np.ma.is_masked(stat_min_max_dict['ax2_stat_max']):
                         if not np.ma.is_masked(ci_max):
                             stat_min_max_dict['ax2_stat_max'] = ci_max
-                    cmasked_ci_thresh_values = np.ma.compressed(
-                        masked_ci_thresh_values
+                    cmasked_ci_valid_hours = np.ma.compressed(
+                        masked_ci_valid_hours
                     )
                     cmasked_model_num_model1_diff_ci_data = np.ma.compressed(
                         masked_model_num_model1_diff_ci_data
@@ -538,15 +529,15 @@ class ThresholdAverage:
                             ci_bar_intvl_widths
                         )
                     )
-                    for thresh_idx in range(len(cmasked_ci_thresh_values)):
-                        thresh = cmasked_ci_thresh_values[thresh_idx]
-                        thresh_ci = (
-                            cmasked_model_num_model1_diff_ci_data[thresh_idx]
+                    for vhr_idx in range(len(cmasked_ci_valid_hours)):
+                        vhr = cmasked_ci_valid_hours[vhr_idx]
+                        vhr_ci = (
+                            cmasked_model_num_model1_diff_ci_data[vhr_idx]
                         )
-                        ax2.bar(thresh, 2*np.absolute(thresh_ci),
-                                bottom=-1*np.absolute(thresh_ci),
-                                width=(cmasked_ci_bar_max_widths[thresh_idx]
-                                       -(cmasked_ci_bar_intvl_widths[thresh_idx]
+                        ax2.bar(vhr, 2*np.absolute(vhr_ci),
+                                bottom=-1*np.absolute(vhr_ci),
+                                width=(cmasked_ci_bar_max_widths[vhr_idx]
+                                       -(cmasked_ci_bar_intvl_widths[vhr_idx]
                                        *model_idx_list.index(model_idx))),
                                 color = 'None',
                                 edgecolor=model_num_plot_settings_dict['color'],
@@ -555,8 +546,9 @@ class ThresholdAverage:
                     self.logger.debug(f"{model_num} ["
                                       +f"{model_num_name},"
                                       +f"{model_num_plot_name}] difference "
-                                      +"from model1 ["
-                                      +f"{self.model_info_dict['model1']['name']},"
+                                      +"from mode1 ["
+                                      +self.model_info_dict['model1']['name']
+                                      +","
                                       +self.model_info_dict['model1']['plot_name']
                                       +"] confidence intervals has no points")
         subplot_num = 1
@@ -566,14 +558,14 @@ class ThresholdAverage:
             preset_y_axis_tick_min = ax.get_yticks()[0]
             preset_y_axis_tick_max = ax.get_yticks()[-1]
             preset_y_axis_tick_inc = ax.get_yticks()[1] - ax.get_yticks()[0]
-            if self.plot_info_dict['stat'] in ['ACC', 'CSI'] and subplot_num == 1:
+            if self.plot_info_dict['stat'] in ['ACC'] and subplot_num == 1:
                 y_axis_tick_inc = 0.1
             else:
                 y_axis_tick_inc = preset_y_axis_tick_inc
             if np.ma.is_masked(stat_min):
                 y_axis_min = preset_y_axis_tick_min
             else:
-                if self.plot_info_dict['stat'] in ['ACC', 'CSI'] and subplot_num == 1:
+                if self.plot_info_dict['stat'] in ['ACC'] and subplot_num == 1:
                     y_axis_min = round(stat_min,1) - y_axis_tick_inc
                 else:
                     y_axis_min = preset_y_axis_tick_min
@@ -582,7 +574,7 @@ class ThresholdAverage:
             if np.ma.is_masked(stat_max):
                 y_axis_max = preset_y_axis_tick_max
             else:
-                if self.plot_info_dict['stat'] in ['ACC', 'CSI'] and subplot_num == 1:
+                if self.plot_info_dict['stat'] in ['ACC'] and subplot_num == 1:
                     y_axis_max = 1
                 else:
                     y_axis_max = preset_y_axis_tick_max + y_axis_tick_inc
@@ -614,11 +606,11 @@ class ThresholdAverage:
             stat_min = stat_min_max_dict['ax1_stat_min']
             stat_max = stat_min_max_dict['ax1_stat_max']
             legend = ax1.legend(
-                bbox_to_anchor=(plot_specs_ta.legend_bbox[0],
-                                plot_specs_ta.legend_bbox[1]),
-                loc = plot_specs_ta.legend_loc,
-                ncol = plot_specs_ta.legend_ncol,
-                fontsize = plot_specs_ta.legend_font_size
+                bbox_to_anchor=(plot_specs_vha.legend_bbox[0],
+                                plot_specs_vha.legend_bbox[1]),
+                loc = plot_specs_vha.legend_loc,
+                ncol = plot_specs_vha.legend_ncol,
+                fontsize = plot_specs_vha.legend_font_size
             )
             plt.draw()
             inv = ax1.transData.inverted()
@@ -638,11 +630,11 @@ class ThresholdAverage:
                     )
                     ax1.set_ylim([y_axis_min, y_axis_max])
                     legend = ax1.legend(
-                        bbox_to_anchor=(plot_specs_ta.legend_bbox[0],
-                                        plot_specs_ta.legend_bbox[1]),
-                        loc = plot_specs_ta.legend_loc,
-                        ncol = plot_specs_ta.legend_ncol,
-                        fontsize = plot_specs_ta.legend_font_size
+                        bbox_to_anchor=(plot_specs_vha.legend_bbox[0],
+                                        plot_specs_vha.legend_bbox[1]),
+                        loc = plot_specs_vha.legend_loc,
+                        ncol = plot_specs_vha.legend_ncol,
+                        fontsize = plot_specs_vha.legend_font_size
                     )
                     plt.draw()
                     inv = ax1.transData.inverted()
@@ -677,7 +669,7 @@ def main():
         'init_hr_start': 'INIT_HR_START',
         'init_hr_end': 'INIT_HR_END',
         'init_hr_inc': 'INIT_HR_INC',
-        'forecast_hour': 'FORECAST_HOUR'
+        'forecast_hours': ['FORECAST_HOURS']
     }
     PLOT_INFO_DICT = {
         'line_type': 'LINE_TYPE',
@@ -689,10 +681,10 @@ def main():
         'interp_points': 'INTERP_POINTS',
         'fcst_var_name': 'FCST_VAR_NAME',
         'fcst_var_level': 'FCST_VAR_LEVEL',
-        'fcst_var_threshs': ['FCST_VAR_THRESHS'],
+        'fcst_var_thresh': 'FCST_VAR_THRESH',
         'obs_var_name': 'OBS_VAR_NAME',
         'obs_var_level': 'OBS_VAR_LEVEL',
-        'obs_var_threshs': ['OBS_VAR_THRESHS'],
+        'obs_var_thresh': 'OBS_VAR_THRESH',
         'fig_name_label': 'FIG_NAME_LABEL',
     }
     MET_INFO_DICT = {
@@ -721,10 +713,10 @@ def main():
     logger_info = f"Log file: {job_logging_file}"
     print(logger_info)
     logger.info(logger_info)
-    p = ThresholdAverage(logger, INPUT_DIR, OUTPUT_DIR, MODEL_INFO_DICT,
+    p = ValidHourAverage(logger, INPUT_DIR, OUTPUT_DIR, MODEL_INFO_DICT,
                          DATE_INFO_DICT, PLOT_INFO_DICT, MET_INFO_DICT,
                          LOGO_DIR)
-    p.make_threshold_average()
+    p.make_valid_hour_average()
 
 if __name__ == "__main__":
     main()
